form_withを用いてフォロー機能を実装

+++ app/models/user.rb
@@ -26,7 +26,7 @@ class User < ApplicationRecord
 
   def unfollow(other_user)
     relationship = self.relationships.find_by(follow_id: other_user.id)
-    relationship.destroy
+    relationship.destroy if relationship
   end
 
   def following?(other_user)
